requirements.txt
huggingface_hub
python-dotenv
torch
os
pathlib
requests
re
transformers
concurrent
pandas
ast
matplotlib
groq
json
time
